+++ diffcustom.sh
@@ -2,20 +2,24 @@
 #take the two masters files in arguments (like A,B or S) in order to compare it
 
 #create a temporary file with only line like : +./toto/tata yy-mm-dd hh:mm:ss or -./ :
-
 diff -u $1 $2 | grep "^[+-].*" | tail -n +3 > temp
 sort -o temp temp
-
-
+#cat temp
+#echo -e "\n"
 #format the temp file in a temp2 file to be useable in a for function :
 
 cut -c 2- temp | cut -f 1 -d " " > temp2
+#cat temp2
 
 #searche for eatch line if the former line is the same
 #if two lines are the same we remove them and write a *toto in diffA :
 
 iminusone="_"
-rm diff$1$2
+
+if [ -f diff$1$2 ]; then
+	rm diff$1$2
+fi
+
 for i in  $(cat temp2)
 do
 	if [ $i = $iminusone ]
